requirements.txt
Django==5.1.1
djangorestframework==3.15.2
channels==4.0.0
channels-redis==4.2.0
celery==5.3.4
redis==5.0.1
Pillow==10.0.1
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0